Remove unnecessary flag. Minor fixes.
